fix: Actions Blade component
