+++ packages/panels/resources/views/components/header/index.blade.php
@@ -34,7 +34,7 @@
     {{ \Filament\Support\Facades\FilamentView::renderHook('panels::page.header.actions.before', scopes: $this->getRenderHookScopes()) }}
 
     @if ($actions)
-        <x-filament-actions::actions
+        <x-filament::actions
             :actions="$actions"
             @class([
                 'shrink-0',
